\name{aln2dat}
\alias{aln2dat}
\title{
Convert ClustalX alignment data to dataframe
}
\description{
Convert ClustalX alignment data to dataframe
}
\usage{
aln2dat(aln)
}
\arguments{
  \item{aln}{
A character string indicating the the aln format file generated by Clustal 
}
}
\details{
None.
}
\value{
Dataframe
}
\references{
None.
}
\author{

Jinlong Zhang \email{jinlongzhang01@gmail.com}

}
\note{
Kress W., Erickson D., Jones F., Swenson N., Perez R., Sanjur O., Bermingham E., Plant DNA barcodes and community phylogeny of a tropical forest dynamics plot in Panama. Proceedings of the National Academy of Sciences of the United States of America. 2009 18621-18626
}

\examples{

dir <- system.file("extdata", package = "phylotools")
setwd(dir)
##aln2dat examle ########
test <- readLines("aln")
aln2dat(test)

}

\keyword{ Algnment }
