\name{read.phylip}
\alias{read.phylip}

\title{
Read phylip file to memory.
}
\description{
Read phylip file to memory.
}
\usage{
read.phylip(fil = NULL)
}
\arguments{
  \item{fil}{
  file path for the input phylip file.
}
}
\details{
The phylip format should be generated by the Clustal X software.
}
\value{
Phylip objected to be used in further analysis.
}
\references{
Kress W., Erickson D., Jones F., Swenson N., Perez R., Sanjur O., Bermingham E., Plant DNA barcodes and community phylogeny of a tropical forest dynamics plot in Panama. Proceedings of the National Academy of Sciences of the United States of America. 2009 18621-18626
}
\author{

Jinlong Zhang \email{jinlongzhang01@gmail.com}

}
\note{
None.
}

\seealso{
See Also as \code{\link{readLines}}
}
\examples{

## read.phylip() examle ########

dir <- system.file("extdata", package = "phylotools")
setwd(dir)
read.phylip("rbcla.phy")

}

\keyword{ phylip }

