\name{resid.tree}
\alias{resid.tree}

\title{
The undeleted terminals of a phylogenetic tree
}
\description{
This is a internal function, and only handling the object that generated by del.tree.tip.
}
\usage{
resid.tree(tree, deltree)
}

\arguments{
  \item{tree}{
A phylogenetic tree
}
  \item{deltree}{
An object that generated by del.tree.tip
}
}

\value{
The phylogenetic tree of the deleted tipes
}

\author{
Jinlong Zhang \email{ jinlongzhang01@gmail.com }
}

\seealso{
\code{\link{del.tree.tip}}
}

\examples{
data(bird.orders)
del <- del.tree.tip(bird.orders,3)
del.tree <- resid.tree(bird.orders, del)
plot(del.tree)
}

