\name{seq2fasta}
\alias{seq2fasta}

\title{
Convert Seq file to fasta format
}
\description{
This function convert .seq file generated by dnaMAN to fasta format.
}
\usage{
seq2fasta(file)
}

\arguments{
  \item{file}{
A character string indicate the ".seq" file name.
}
}
\details{
Seq file is a DNA format generated by dnaMAN, often the exported sequences assembly results. Sometimes these files will have to be converted to fasta format in order to carry on further analysis. 
}
\value{
fasta file.
}
\references{
To be added 
}
\author{
Jinlong Zhang \email{jinlongzhang01@gmail.com}
}

\examples{

dir <- system.file("extdata", package = "phylotools")
seq2fasta(paste(dir,"/DNAman.seq", sep = ""))

}
