\name{split_dat}
\alias{split_dat}

\title{
grouping the data frame containing sequences and names and generate fasta file
}
\description{
Splite the data frame of sequences based on the reference table of grouping.
}
\usage{
split_dat(dat, ref_table)
}

\arguments{
  \item{dat}{
data frame generated by \code{\link{read.phylip}} or \code{\link{read.fasta}}
}
  \item{ref_table}{
data frame with first column for the name of the sequence, second column for the group the sequence belongs to.
}
}
\details{
Each group of sequences will be saved to a fasta file. Sequences not included in the ref_table will be saved in "Ungrouped.fasta"
}
\value{
This is a subroutine, there is no return value.
}
\references{
\url{http://www.genomatix.de/online_help/help/sequence_formats.html}
}
\author{
Jinlong Zhang <jinlongzhang01@gmail.com>
}

\seealso{
\code{\link{rename.fasta}}
}
\examples{

  cat(
  ">seq_1",   "--TTACAAATTGACTTATTATA",
  ">seq_2",   "GATTACAAATTGACTTATTATA",
  ">seq_3",   "GATTACAAATTGACTTATTATA",
  ">seq_5",   "GATTACAAATTGACTTATTATA",
  ">seq_8",   "GATTACAAATTGACTTATTATA",
  ">seq_10",  "---TACAAATTGAATTATTATA",
  ">seq_11",  "--TTACAAATTGACTTATTATA",
  ">seq_12",  "GATTACAAATTGACTTATTATA",
  ">seq_13",  "GATTACAAATTGACTTATTATA",
  ">seq_15",  "GATTACAAATTGACTTATTATA",
  ">seq_16",  "GATTACAAATTGACTTATTATA",
  ">seq_17",  "---TACAAATTGAATTATTATA",
  file = "trnH.fasta", sep = "\n")

sequence_name <- get.fasta.name("trnH.fasta")
sequence_group <- c("group1","group1","group1","group1","group1",
"group2","group2","group2","group3","group3","group3","group3")
group <- data.frame(sequence_name, sequence_group)

fasta <- read.fasta("trnH.fasta")
split_dat(fasta, group)

}

\keyword{ fasta }
