\name{phylotools-package}
\alias{phylotools-package}
\alias{phylotools}
\docType{package}
\title{
Phylogenetic tools for building PHYLIP supermatrix and more
}
\description{
A collection of a few functions for handling DNA-barcoding sequences,
building PHYLIP supermatrix for RAxML etc.
}
\details{
\tabular{ll}{
Package: \tab phylotools\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2015-05-26\cr
License: \tab GLP-2\cr
LazyLoad: \tab yes\cr
}

}
\author{

Jinlong Zhang

Maintainer: Jinlong Zhang <jinlongzhang01@gmail.com>

}

\keyword{ package }

